+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Revision history for monoid-map
 
+## UNRELEASED
+
+* Add support for DecidablyEmpty
+* Get rid of invalid Functor, Traversable, and Filterable instances; create TypeError for Functor instance
+
 ## v0.1.1.0
 
 * Remove dependency on `appendmap`